add randomization to scenarios
