+++ pubspec.yaml
@@ -37,6 +37,7 @@ dependencies:
   envied:
   url_launcher:
   flutter_bloc:
+  dart_random_choice:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
